freerdp: fix type build warnings

+++ winpr/libwinpr/synch/wait.c
@@ -457,7 +457,7 @@ DWORD WaitForSingleObject(HANDLE hHandle, DWORD dwMilliseconds)
 	}
 	else
 	{
-		fprintf(stderr, "WaitForSingleObject: unknown handle type %d\n", Type);
+		fprintf(stderr, "WaitForSingleObject: unknown handle type %d\n", (int) Type);
 	}
 
 	return WAIT_OBJECT_0;
@@ -549,7 +549,7 @@ DWORD WaitForMultipleObjects(DWORD nCount, const HANDLE* lpHandles, BOOL bWaitAl
 		}
 		else
 		{
-			fprintf(stderr, "WaitForMultipleObjects: unknown handle type %d\n", Type);
+			fprintf(stderr, "WaitForMultipleObjects: unknown handle type %d\n", (int) Type);
 			return WAIT_FAILED;
 		}
